+++ p2n/ops/decoder.py
@@ -2,11 +2,13 @@
 # (c) 2017 The Patent2Net Developers
 import re
 import logging
+from copy import deepcopy
 from collections import OrderedDict
 from jsonpointer import JsonPointer, JsonPointerException
 from p2n.util import to_list
 
 logger = logging.getLogger(__name__)
+
 
 class OPSExchangeDocumentDecoder:
     """
@@ -45,9 +47,14 @@ class OPSExchangeDocumentDecoder:
                 else:
                     doc_number = document_id['country']['$'] + document_id['doc-number']['$'] + document_id['kind']['$']
                 date = document_id.get('date', {}).get('$')
-                date = date and '-'.join([date[:4], date[4:6], date[6:8]])
+                date = cls.decode_date(date)
                 return doc_number, date
         return None, None
+
+    @staticmethod
+    def decode_date(date):
+        date = date and '-'.join([date[:4], date[4:6], date[6:8]])
+        return date
 
     @classmethod
     def titles(cls, data):
@@ -147,7 +154,10 @@ class OPSExchangeDocumentDecoder:
     @classmethod
     def classifications_ipcr(cls, data):
 
-        entries = to_list(cls.pointer_ipcr.resolve(data))
+        try:
+            entries = to_list(cls.pointer_ipcr.resolve(data))
+        except JsonPointerException:
+            return []
 
         entries = map(lambda entry: entry['text']['$'], entries)
         entries = map(lambda entry: entry[:15].replace(' ', ''), entries)
@@ -195,3 +205,200 @@ class OPSExchangeDocumentDecoder:
                 classifications[scheme].append(entry)
 
         return classifications
+
+
+
+class OPSRegisterDocumentDecoder:
+    """
+    Functions for decoding data from raw JSON OPS register documents.
+    """
+
+    pointer_status = JsonPointer('/reg:register-document/@status')
+    pointer_designated_states = JsonPointer('/reg:register-document/reg:bibliographic-data/reg:designation-of-states')
+    pointer_applicants = JsonPointer('/reg:register-document/reg:bibliographic-data/reg:parties/reg:applicants')
+    pointer_inventors = JsonPointer('/reg:register-document/reg:bibliographic-data/reg:parties/reg:inventors')
+    pointer_agents = JsonPointer('/reg:register-document/reg:bibliographic-data/reg:parties/reg:agents')
+    pointer_actions = JsonPointer('/reg:register-document/reg:bibliographic-data/reg:dates-rights-effective')
+    pointer_filing_language = JsonPointer('/reg:register-document/reg:bibliographic-data/reg:language-of-filing/$')
+
+    """
+    TODO:
+
+    # EP88402018P
+    u'reg:dates-rights-effective': {u'reg:first-examination-report-despatched': {u'reg:date': {u'$': u'19901213'}},
+                                    u'reg:request-for-examination': {u'@change-gazette-num': u'1989/41',
+                                                                     u'reg:date': {u'$': u'19890628'}}},
+
+    # EP86400204P
+    u'reg:opposition-data': {u'@change-date': u'19890321',
+                             u'@change-gazette-num': u'1989/19',
+                             u'reg:opposition-not-filed': {u'reg:date': {u'$': u'19890221'}}},
+
+    # EP06823909P
+    u'reg:date-application-deemed-withdrawn': {u'@change-gazette-num': u'2009/11',
+                                               u'reg:date': {u'$': u'20080909'}},
+
+    # EP10806019P
+    u'reg:date-application-withdrawn-by-applicant': {u'@change-gazette-num': u'2012/35'},
+
+
+    # EP08836401P
+    u'reg:term-of-grant': [{u'@change-date': u'20140718',
+                            u'@change-gazette-num': u'2014/34',
+                            u'reg:lapsed-in-country': [{u'reg:country': {u'$': u'HU'},
+                                                        u'reg:date': {u'$': u'20080709'}},
+                                                       {u'reg:country': {u'$': u'AT'},
+                                                        u'reg:date': {u'$': u'20120418'}},
+
+    # EP16202765P
+    u'reg:related-documents': {u'reg:division': {u'reg:relation': {u'reg:child-doc': {u'reg:document-id': {u'reg:country': {u'$': u''},
+           u'reg:doc-number': {u'$': u''}}},
+
+
+    # EP14879896P
+    u'reg:office-specific-bib-data':
+
+
+    # EP14879896P
+    u'reg:ep-patent-statuses': {u'reg:ep-patent-status': [{u'$': u'The application is deemed to be withdrawn',
+       u'@change-date': u'20170714',
+       u'@status-code': u'10'},
+      {u'$': u'The international publication has been made',
+       u'@change-date': u'20161115',
+       u'@status-code': u'17'}]}}}
+
+
+    """
+
+    @classmethod
+    def status(cls, data):
+        """
+        Decode register status.
+        """
+        return cls.pointer_status.resolve(data)
+
+
+    @classmethod
+    def filing_language(cls, data):
+        """
+        Decode register status.
+        """
+        return cls.pointer_filing_language.resolve(data)
+
+
+    @classmethod
+    def designated_states(cls, data):
+        """
+        Decode designated states from register document.
+
+        # TODO: Multiple designated states entries. e.g. EP16202765P
+        """
+
+        entries = cls.pointer_designated_states.resolve(data)
+
+        countries_reference = JsonPointer('/reg:designation-pct/reg:regional/reg:country')
+        countries_raw = to_list(countries_reference.resolve(entries))
+
+        states = []
+        for country_raw in countries_raw:
+            country = country_raw['$']
+            states.append(country)
+
+        return states
+
+
+    @classmethod
+    def parties(cls, partylist):
+        """
+        Decode list of applicants or inventors.
+        """
+        #print 'partylist:', partylist
+        entries = []
+        for party in partylist:
+
+            addressbook = party['reg:addressbook']
+
+            entry = OrderedDict()
+            entry['name'] = addressbook['reg:name']['$']
+            entry['country'] = addressbook['reg:address']['reg:country']['$']
+            entry['city'] = addressbook['reg:address']['reg:address-2']['$']
+            entry['address'] = addressbook['reg:address']['reg:address-1']['$']
+
+            entries.append(entry)
+
+        return entries
+
+    @classmethod
+    def parties_history(cls, nodes, node_name):
+
+        # Collect entries over time
+        history = []
+        for node in nodes:
+            entry = OrderedDict()
+            entry['change_date'] = OPSExchangeDocumentDecoder.decode_date(node['@change-date'])
+            entry['change_gazette'] = node['@change-gazette-num']
+            entry['items'] = cls.parties(to_list(node[node_name]))
+            history.append(entry)
+
+        # Deduplicate entries. Sometimes, duplicate entries are in the history list,
+        # one with 'change_gazette' == 'N/P' and another one with a real value, e.g. '1986/34'.
+        # We want to choose the entry with the real value and suppress to other one,
+        # but only if all the bibliographic data are equal.
+        deduplicated = []
+        real = []
+        for entry in history:
+            entry_dup = deepcopy(entry)
+            del entry_dup['change_gazette']
+
+            if deduplicated:
+                overwrite = (real[-1]['change_gazette'] == 'N/P' and real[-1]['change_date'] == entry['change_date']) and (deduplicated[-1] == entry_dup)
+                if overwrite:
+                    real.pop()
+                    deduplicated.pop()
+
+            if entry_dup not in deduplicated:
+                deduplicated.append(entry_dup)
+                real.append(entry)
+
+        return real
+
+    @classmethod
+    def applicants(cls, data):
+        try:
+            nodes = to_list(cls.pointer_applicants.resolve(data))
+        except JsonPointerException:
+            return []
+
+        return cls.parties_history(nodes, 'reg:applicant')
+
+    @classmethod
+    def inventors(cls, data):
+        try:
+            nodes = to_list(cls.pointer_inventors.resolve(data))
+        except JsonPointerException:
+            return []
+
+        return cls.parties_history(nodes, 'reg:inventor')
+
+    @classmethod
+    def agents(cls, data):
+        try:
+            nodes = to_list(cls.pointer_agents.resolve(data))
+        except JsonPointerException:
+            return []
+
+        return cls.parties_history(nodes, 'reg:agent')
+
+    @classmethod
+    def actions(cls, data):
+        try:
+            actions = []
+            for key, value in cls.pointer_actions.resolve(data).iteritems():
+                entry = {
+                    'name': key.replace('reg:', ''),
+                    'date': OPSExchangeDocumentDecoder.decode_date(value['reg:date']['$']),
+                }
+                actions.append(entry)
+            return actions
+        except JsonPointerException:
+            return []